% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bcf_get_n_samples}
\alias{bcf_get_n_samples}
\title{Get the number of samples in a BCF/VCF file}
\usage{
bcf_get_n_samples(filename)
}
\arguments{
\item{filename}{Path to a BCF/VCF file}
}
\value{
Number of samples
}
\description{
Get the number of samples in a BCF/VCF file
}
