% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bcf_get_sample_names}
\alias{bcf_get_sample_names}
\title{Get sample names from a BCF/VCF file}
\usage{
bcf_get_sample_names(filename)
}
\arguments{
\item{filename}{Path to a BCF/VCF file}
}
\value{
Character vector of sample names
}
\description{
Get sample names from a BCF/VCF file
}
