% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltrepBcfRecords.R
\name{BcfGetNSamples}
\alias{BcfGetNSamples}
\title{Get the number of samples in a BCF/VCF file}
\usage{
BcfGetNSamples(filename)
}
\arguments{
\item{filename}{Path to a BCF/VCF file}
}
\value{
Number of samples
}
\description{
Get the number of samples in a BCF/VCF file
}
