% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltrepBcfRecords.R
\name{BcfGetSampleNames}
\alias{BcfGetSampleNames}
\title{Get sample names from a BCF/VCF file}
\usage{
BcfGetSampleNames(filename)
}
\arguments{
\item{filename}{Path to a BCF/VCF file}
}
\value{
Character vector of sample names
}
\description{
Get sample names from a BCF/VCF file
}
