\name{bcflib-package}
\alias{bcflib-package}
\alias{bcflib}
\docType{package}
\title{\packageTitle{bcflib}}
\description{\packageDescription{bcflib}}
\section{Package Content}{\packageIndices{bcflib}}
\author{\packageAuthor{bcflib}}
\section{Maintainer}{\packageMaintainer{bcflib}}
\keyword{package}
