% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BcftoolsMerge.R
\name{bcftoolsMergeDirect}
\alias{bcftoolsMergeDirect}
\title{Run BCFtools Merge with Direct Parameters}
\usage{
bcftoolsMergeDirect(
  output_fname = "-",
  input_files = characterVector(),
  output_type = "v",
  compression_level = -1L,
  n_threads = 0L,
  missing_to_ref = FALSE,
  force_samples = FALSE,
  header_only = FALSE,
  header_fname = "",
  regions = "",
  info_rules = "",
  missing_rules = "",
  collapse_type = "both",
  trim_star_allele = FALSE,
  trim_star_allele_all = FALSE,
  local_alleles = 0L,
  force_single = FALSE,
  filter_logic = "",
  gvcf_fai_fname = "",
  record_cmd_line = TRUE,
  write_index = FALSE
)
}
\arguments{
\item{output_fname}{Output filename ["-"] (stdout)}

\item{input_files}{Character vector of input VCF/BCF files}

\item{output_type}{Output type: "v" (VCF), "z" (compressed VCF), "b" (BCF), "u" (uncompressed BCF) ["v"]}

\item{compression_level}{Compression level for output (0-9, -1 for default) [-1]}

\item{n_threads}{Number of extra threads to use [0]}

\item{missing_to_ref}{Assume genotypes at missing sites are reference [FALSE]}

\item{force_samples}{Resolve duplicate sample names [FALSE]}

\item{header_only}{Print only the merged header and exit [FALSE]}

\item{header_fname}{Use the provided header [NULL]}

\item{regions}{Restrict to comma-separated list of regions [NULL]}

\item{info_rules}{Rules for merging INFO fields (e.g., "DP:sum,DP4:sum") [NULL]}

\item{missing_rules}{Rules for replacing missing values (e.g., "PL:.,AD:0") [NULL]}

\item{collapse_type}{How to handle multiallelic records: "snps", "indels", "both", "all", "none", "snp-ins-del", "id" ["both"]}

\item{trim_star_allele}{Remove unobserved alleles if site is variant [FALSE]}

\item{trim_star_allele_all}{Remove unobserved alleles at all sites [FALSE]}

\item{local_alleles}{If more than this number of alleles are encountered, use local alleles (0=unlimited) [0]}

\item{force_single}{Run even if there is only one input file [FALSE]}

\item{filter_logic}{Filter logic: "+" (apply all filters), "x" (remove filters if one is PASS) [NULL]}

\item{gvcf_fai_fname}{Reference FASTA for gVCF block merging [NULL]}

\item{record_cmd_line}{Append version and command line to header [TRUE]}

\item{write_index}{Automatically index the output file [FALSE]}
}
\value{
Integer exit code from bcftools merge (0 on success).
}
\description{
Run bcftools merge with parameters passed directly through R, avoiding the need to construct command-line arguments.
}
\examples{
\dontrun{
bcftoolsMergeDirect(
    output_fname = "merged.vcf.gz",
    input_files = c("file1.vcf.gz", "file2.vcf.gz"),
    output_type = "z",
    n_threads = 4
)
}

}
