% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{faidx_fetch_region}
\alias{faidx_fetch_region}
\title{Fetch a sequence region from a FASTA file using htslib faidx}
\usage{
faidx_fetch_region(fasta_path, seqname, start, end)
}
\arguments{
\item{fasta_path}{Path to the FASTA file}

\item{seqname}{Chromosome/contig name (e.g., "chr1")}

\item{start}{1-based start position (inclusive)}

\item{end}{1-based end position (inclusive)}
}
\value{
Character vector of the fetched sequence
}
\description{
Fetch a sequence region from a FASTA file using htslib faidx
}
