% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{faidx_index_fasta}
\alias{faidx_index_fasta}
\title{Index a FASTA file using htslib faidx and return the index path}
\usage{
faidx_index_fasta(fasta_path)
}
\arguments{
\item{fasta_path}{Path to the FASTA file}
}
\value{
Path to the generated .fai index file
}
\description{
Index a FASTA file using htslib faidx and return the index path
}
