% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBcftoolsVersion}
\alias{getBcftoolsVersion}
\title{Get BCFtools version}
\usage{
getBcftoolsVersion()
}
\value{
Character string of BCFtools version.
}
\description{
Returns the version string of the linked BCFtools library.
}
