% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heterozygosity}
\alias{heterozygosity}
\title{Calculate per-sample heterozygosity from a VCF/BCF file}
\usage{
heterozygosity(vcffile, region = "", samples = "")
}
\arguments{
\item{vcffile}{Path to the VCF/BCF file.}

\item{region}{Optional genomic region (e.g., "chr1:1000-2000").}

\item{samples}{Optional comma-separated list of sample names.}
}
\value{
Integer vector with the count of heterozygous genotypes for each sample.
}
\description{
This function computes the number of heterozygous genotypes for each sample in a given VCF/BCF file.
Only diploid genotypes are considered. The function can be restricted to a genomic region and/or a subset of samples.
}
