% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BcftoolsMerge.R
\name{bcftoolsMerge}
\alias{bcftoolsMerge}
\title{Run BCFtools Merge}
\usage{
bcftoolsMerge(args)
}
\arguments{
\item{args}{Character vector of arguments to pass to bcftools merge.}
}
\value{
Integer exit code from bcftools merge (0 on success).
}
\description{
Run bcftools merge to combine multiple VCF/BCF files to create one multi-sample file.
}
\examples{
\dontrun{
bcftoolsMerge(c("file1.vcf.gz", "file2.vcf.gz", "-o", "merged.vcf.gz", "-O", "z"))
}

}
